[Fuzzing] Move all storage related fuzzers into the same file.
