+++ testsuite/libra-fuzzer/src/fuzz_targets.rs
@@ -9,6 +9,7 @@ use std::{collections::BTreeMap, env};
 // List fuzz target modules here.
 mod compiled_module;
 mod consensus_proposal;
+mod execute_and_commit_chunk;
 mod inbound_rpc_protocol;
 mod inner_signed_transaction;
 mod json_rpc_service;
@@ -16,9 +17,7 @@ mod language_transaction_execution;
 mod mempool;
 mod network_noise;
 mod state_sync_msg;
-//mod storage_save_blocks;
-mod execute_and_commit_chunk;
-mod storage_schema_decode;
+mod storage;
 mod vm_value;
 
 static ALL_TARGETS: Lazy<BTreeMap<&'static str, Box<dyn FuzzTargetImpl>>> = Lazy::new(|| {
@@ -36,7 +35,7 @@ static ALL_TARGETS: Lazy<BTreeMap<&'static str, Box<dyn FuzzTargetImpl>>> = Lazy
         Box::new(state_sync_msg::StateSyncMsg::default()),
         Box::new(language_transaction_execution::LanguageTransactionExecution::default()),
         //        Box::new(storage_save_blocks::StorageSaveBlocks::default()),
-        Box::new(storage_schema_decode::StorageSchemaDecode::default()),
+        Box::new(storage::StorageSchemaDecode::default()),
         Box::new(vm_value::ValueTarget::default()),
         Box::new(execute_and_commit_chunk::ExecuteAndCommitChunk::default()),
     ];
